Fixed Data Loading Error
Data now loads properly as far as TTDRS and ClaimSummary. TTDRS was not
initially inserted as a programming oversight, but now is implemented to
insert when ClaimSummary is created. If neither returns from the DB both
are set to null and accounted for. Also, claimantList is now initialized
with first RO selected.