Added Delete Paycheck button; Fixing paycheck InjDate overlap
Added Button and method funtionality to delete a single Paycheck from
claim. Discovered bug where paychecks that overlap Date Injured do not
compute right. progress made on this but not finished. Next step is
fixing SharedListSelectionHandler as it attempts to access TPDRSumm
before it is instantiated. Root of problem and fix is known, detailed in
source comments.