New TPDPC class; altered calculated pay/owed strat
Created new class "TPDPaycheck" to include a BigDecimal for the
calculated WorkComp payment for that period. After remembering that the
purpose of this application is to help people keep track of their work
comp payments to make sure they are properly compensated for lost wages,
I was reminded that Work Comp supplemental checks are often highly
inconsistent with the frequency in which they are issued. Therefore,
having a 1:1 or even 2:1 mapping of WCPCs to PCs is impractical and
would require more intensive date evaluation. But even then, my own
WCPCs from my claim never specified what period the checks were for. So
in this instance, it becomes necessary to compare the totals of WorkComp
calculated payments for a TPD work period and the total actual
supplemental WCPC amounts. This way it is at least possible to tell,
through manual evaluation of the data, from which TPD pay period that
any differences in totals likely come from if payment must be disputed.

Also, Added some minor code optimizations for calculating totals and
navigating through ArrayLists.