Fixed TPDPC Class type instancing; implm. pc split
Fixed instancing issues of TPDPC classes in various method passings.
Also, added method to SLC to split a paycheck if the Menu method finds
that the dates overlap the end of the prior week to injury date. Splits
PCs into a priorWage pc and a TPDPC. Adds TPDPC to TPDRSUMM and
initializes the TPDRSUMM.