Added selectedROEnabler() method in Menu
selectedROEnabler() in the Menu class does the same thing as the
ListSelectionEvenHandler (which now calls the new method instead of
implementing the logic itself). This fixes the issue where the values
and buttons of the selected RO weren't updated autmatically after
completing the method withing the button ActionEvent. The previous
approach (setting selected index to the same index) did not fire the
valueChanged method of the LSEH and the new approach fixes that
(presumably currently). This was done in place of having to call
EventQueue to fire a pseudo valueChanged event call.