Implemented TTD No PP Date ANP Computation
Implemented features that prompt user if they know the PPS and PPE dates
for the TTD WCPC. If not, those dates are both set to the Epoch and the
ANP for TTD is then calculated by total days not in TPD, but with no
Late Payment adjustments initially. Should correctly display, but
untested. Testing this is the next TODO.