Fixed WCPC SortedMap value insertion
Would not add multuple WCPCs with unknown PayPeriod dates to be
displayed in "Delete Paychecks".

This is because the dates were set to the epoch so each PPS date
compares to 0, always overwriting the previous value in SortedMap. Fixed
WC_COMPARATOR final field value to appropriately compare payment dates
instead.