Calculations and inits fixed; TPD WC Pay doubled
Fixed computational errors among a number of calculations which resulted
from TPDPaycheck.computeWCCalcPay() not getting called during or
imidiately after initializing. Added a sLC to TPDPC to simplify this
process.

Next TODO is to to identify why the TPD WC PayToDate amonut is excactly
double of what it should be even though it is called from the same super
class RS method as the TTD method which does not return a doubled
amount.

Also, planned TODO to implement calculation logging so that the math
behind the computations is displayed.