Fixed doubling of WC Paid
Found where TPD WCPCs were getting inserted twice, but was not shown by
the delete method because duplicates are replaced since it uses a
hashmap. Removed cause of double insertion and deleted existing
duplicate inserts.
Next TODO: need to find why ANP value for TPD and total is not being
recalculated with this change.