Impl LD; Computations return correct and consistent
Implemented a button to set the Light Duty Start Date to accuately
divide the TTD and TPD periods so that the full number of days for TTD
could be accounted for. This allowed the aNP and totals owed to be
computed correctly and consistently.  Also added Light Duty Date to RO
and created appropriate methods for insertion and selection from DB.
Also created implementation to automatically change a TPDPCs PPS date to
the LDStart if the PPS was before the LDStart.

All that remains TODO is creating a logging method to display the math
used to make all of the computations. This is so that legal consumers
can verify the computations.