TPDRS Recovered; LICENSE added, README updated
Was able to revover TPDReimbursementSummary which had been lost in a
system crash. Found using Recuva using a deep scan with "search for a
specific string within file contents..." using the class name as the
string. The entire .java file had previously been overwritten with char
"\u000" due to an unknown error when the crash occurred. Contents were
recovered and appears to have fixed the class. I cannot remember
previously if I had finished the MathLogging for WCPC, but it is
currently in an unfinished state. Will finish that next.

Added a LICENSE.md to the repository which licenses this repository for
research use only and declares it  and all contents non-distributable
without permission from myself.  Updated the README to note the addition
of this License.