Database implementation for MathLogs
Created table for MATH_LOGS and implmented insertion and selection for
all MATH_LOGS. In reality, this should only be necessary for paychecks
that have been trimmed and have had recalculated monetary amounts as
such. This required getting keys for a paycheck each time the log is
updated or selected without having any other sort of specific identifier
for a PC mathLog.  Still need to implement the mathLogging for PC, WCPC,
and TPDPC.